New AI Features (#143)
# Description
Adds some cool new stuff to the AI, including random screens and laws on
startup.

# Media
See diff bot.


# Changelog

:cl:
- add: Nanotrasen has slightly debungled the AI from its semi-lucid,
ludonarrative dissonance realm from which it resides and dragged it
kicking and screaming into the world of story. In other terms, the AI is
prettier, and NT is testing out a series of new laws that may make their
way onto a station near you! Be sure to mark down your satisfaction
level on the employee report, if you get that far.