I mean, it freakin works :P What else can ya ask for.
There needs to be a proper way to specify skill requirements in YAML, likely a serializable struct. There's a basis for that commented out in the main System at the moment but I wanted to test if it actually worked.
Other than that, I suppose it's just a matter of refining, helper functions, and of course, implementation.