Lotsa very temp work.
I'm handing this over to Death at this point. Important notes below.

The UI doesn't automatically update (just needs to detect changes server side and UpdateUi to any AiEye, not hard and I can do it later no problem).

The UI is now networked properly (you complained so much :P) making the AiCameraComponent effectively obsolete!
There's still Holo stuff in this pr using it so no touching yet, but that's good, it was hacky.

The big thing right now is that the detection of clicks on the map is a bit of a mess.
It's all written into the NavMapControl itself right now (obviously not ideal) but I see no reason the current code couldn't be moved to the BUI since it's *mostly* public properties.
_offset is private, but some small modifications is better than what I'm doing right now. Regardless, it's easier to dev in there so we'll keep doing it until we have something working.
Speaking of which, it only *kinda* works right now. It's a prototype, and it kinda looks like it works for a bit, but the map doesn't offset properly and even if you line it up at a point
it doesn't scale right.

Note for Death- I'm handing it off to you for two reasons- One, this is a collaborative effort and I've lost enough brain energy to it. Two- holy shit Death this old UI code. I absolutely cannot fathom.
It'll be your job to clean this up into something usable later, since you made the original mess. I made some of it better (probably some of it worse), and I'll obviously help out more, but yeah.
You gotta go over it.